gemma-3-api/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings

llama-cpp-python
huggingface-hub==0.24.1
websockets
openai>=1.30.0
httpx>=0.27.0
python-multipart
openai-whisper
soundfile
prometheus-client
pytest
pytest-asyncio
